// src/web/src/api/generated/models/ActivityRecentComparisonResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ActivityRecentComparisonResponse = {
    windowDays: number | string;
    sampleSize: number | string;
    minDistanceM: number | string | null;
    maxDistanceM: number | string | null;
    medianPaceSecondsPerKm: number | string | null;
    medianHeartRateBpm: number | string | null;
};